Test changes in entity ID -> entity set ID associations for '{add,delete}Entity'.
Among the side-effects of 'addEntity' and 'removeEntity' is updating
the association from the entity ID to the entity set ID:

  addEntity eid val esid

will make 'esid' appear in the result of

  memberOfSets eid

if it isn't already there. Similarly,

  deleteEntity eid esid

will remove 'esid' from the result of

  memberOfSets eid.

This enhances 'prop_addEntity' and 'prop_deleteEntity' to check for
this.